+++ host/class/uvc/usb_host_uvc/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Added `uvc_host_stream_format_get()` function that returns current stream's format
 - Added `uvc_host_buf_info_get()` function for esp_video binding
+- Added option to request default FPS by setting FPS = 0
 
 ## 2.2.0